tweaks from beta testing

+++ GymRats/feature/workout/ImportWorkoutViewController.swift
@@ -12,7 +12,7 @@ import RxSwift
 import RxDataSources
 
 typealias ImportWorkoutSection = SectionModel<Void, ImportWorkoutRow>
-typealias StepCount = Double
+typealias StepCount = Int
 
 protocol ImportWorkoutViewControllerDelegate: class {
   func importWorkoutViewController(_ importWorkoutViewController: ImportWorkoutViewController, imported workout: HKWorkout)
